requirements.txt
ujson
tqdm
pillow
pymupdf
pdfminer.six
sentence-transformers
faiss-cpu
rank-bm25
transformers
torch
easyocr
numpy
nltk
pyyaml
requests

